refactor(app): extract pollers/watchdog to internal pkgs
- IssuesFetcher → internal/poll/issues.go
- RenovateHandler → internal/poll/renovate.go
- TodoistHandler → internal/poll/todoist.go (TaskCreator func
  replaces *TaskService dep; DomainHandler inlined)
- agentWatchdogLoop → internal/watchdog/agent.go (Watchdog struct,
  emit func replaces runtime.EventsEmit; wg passed via constructor)

App struct loses issuesFetcher field; todoistHandler/renovateHandler
now typed as *poll.T. IntegrationService updated to match.
